// src/server/api/models/insurance/models/Policy.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { InsuranceProductType } from './InsuranceProductType';
import type { LocalDate } from './LocalDate';
import type { LocalDateTime } from './LocalDateTime';
import type { PolicyPaymentTerms } from './PolicyPaymentTerms';
import type { PolicyRiskType } from './PolicyRiskType';
import type { PolicyStatus } from './PolicyStatus';
import type { UUID } from './UUID';

export type Policy = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The client ID from the crm service
     */
    clientId: UUID;
    /**
     * The client description
     */
    clientLabel: string;
    /**
     * The client code from CRM
     */
    clientCode: string;
    /**
     * A unique policy number
     */
    policyNo: string;
    /**
     * A policy number generated by the insurance partner
     */
    refPolicyNo: string;
    /**
     * The policy description
     */
    policyLabel: string;
    /**
     * The policy currency
     */
    currency: string;
    /**
     * Notes related to the record
     */
    notes?: string | null;
    /**
     * The status of the policy
     */
    status: PolicyStatus;
    /**
     * The type of risk insured
     */
    riskType: PolicyRiskType;
    /**
     * Indicates the payment terms for the premium
     */
    paymentTerms: PolicyPaymentTerms;
    /**
     * The premium charged for the policy
     */
    premium: number | null;
    /**
     * The commission charged for the policy
     */
    commission: number | null;
    /**
     * The tax charged for the policy
     */
    tax: number | null;
    /**
     * The total amount charged for the policy
     */
    total: number | null;
    /**
     * The value insured
     */
    sumInsured: number | null;
    /**
     * The amount due in each payment period
     */
    scheduledAmount: number | null;
    /**
     * Indicates the start date for the policy
     */
    startDate: LocalDate;
    /**
     * Indicates the end date for the policy
     */
    endDate: LocalDate;
    /**
     * The policy product ID
     */
    productId: UUID | null;
    /**
     * The policy product label
     */
    productLabel: string | null;
    /**
     * The type of the policy
     */
    productType: InsuranceProductType;
    branchId: UUID;
    branchLabel: string;
    partnerId?: UUID | null;
    partnerLabel?: string | null;
};